Tag "in-progress" items for taskboard visualization

The viz_taskboard() function creates a three-column layout of entities that are not started, in progress, or done. Objects are classified as done when they have a state of "closed". Object are classified as "To-Do" when they are neither "Closed" or "In Progress". However, what constistutes "In Progress" is user and project dependent. Thus, these functions let users specify what they mean.

is_labeled()

is_labeled_with(label)

is_assigned()

is_assigned_to(login)

is_in_a_milestone()

is_in_milestone(number)

is_created_before(created_date)

is_part_closed()

is_due()

is_due_before(due_date)

Arguments

label
Label name as character

login
User login as character

number
Milestone number

created_date
Date as character in "YYYY-MM-DD" format

due_date
Date as character in "YYYY-MM-DD" format

Value

Function to be passed as in_progress_when argument in viz_taskboard()

Details

General options:

is_created_before: Was created before a user-specified data (as "YYYY-MM-DD" character string)

Issue-specific options:

is_labeled_with: User-specified label (as character string) exists

is_assigned: Has been assigned to anyone

is_assigned_to: Has been assigned to specific user-specified login (as character string)

is_in_a_milestone: Has been put into any milestone

is_in_milestone: Has been put into a specific milestone

Milestone-specific options:

is_part_closed: Has any of its issues closed

is_due: Has a due date

is_due_before: Has a due data by or before a user-specified date (as "YYYY-MM-DD" character string)

Examples

# NOT RUN {
viz_taskboard(issues, in_progress_when = is_labeled_with('in-progress'))
viz_taskboard(milestones, in_progress_when = is_created_before('2018-12-31'))
viz_taskboard(issues, in_progress_when = is_in_milestone())
report_taskboard(issues, in_progress_when = is_labeled_with('in-progress'))
report_taskboard(milestones, in_progress_when = is_created_before('2018-12-31'))
report_taskboard(issues, in_progress_when = is_in_milestone())
# }