Visualize Agile-style task board of item status

Produces three column task board showing any relevant objects (typically issues or milestones) as "Not Started", "In Progress", or "Done".

viz_taskboard(data, in_progress_when, str_wrap_width = 30,
  text_size = 3)

Arguments

data
Dataset, such as those representing issues or milestones (i.e. from parse_issues()). Must have state variable.

in_progress_when
Function with parameter data that returns Boolean vector. Generally, one of the taskboard helper functions. See ?taskboard_helpers for details.

str_wrap_width
Number of characters before text of issue title begins to wrap

text_size
Text size

Value

ggplot object

Details

The following logic is used to determine the status of each issue:

Done: Items with a state of "closed"

In Progress: Custom logic via in_progress_when. See ?taskboard_helpers for details.

Not Started: Default case for items neither In Progress or Closed

See also

Examples

# NOT RUN {
issues <- get_issues(myrepo, milestone = 1) %>% parse_issues()
viz_taskboard(issues, in_progress_when = is_labeled_with('in-progress'))
viz_taskboard(issues, in_progress_when = is_in_a_milestone())
# }