Parse issue events from get_issues_events

This function convert list output returned by get into a dataframe. Due to the diverse fields for different types of events, many fields in the dataframe may be NA.

parse_issue_events(res)

Arguments

res
List returned by corresponding get_ function

Value

tibble datasets with one record / issue-event

Details

Currently, the following event types are unsupported (with regard to processing all of their fields) due to their additional bulk and limited utility with respect to this packages functionality. Please file an issue if you disagree:

"(removed_from/moved_columns_in/added_to)_project"Since this package has limited value with GitHub projects

"converted_note_to_issue"Since issue lineage is not a key concern

"head_ref_(deleted/restored)"Since future support for pull requests would likely be handled separately

"merged"Same justification as head_ref

"review_(requested/dismissed/request_removed)Same justification as head_ref

See also

Examples

# NOT RUN {
myrepo <- create_repo_ref('emilyriederer', 'myrepo')
events_res <- get_issue_events(myrepo, number = 1)
events <- parse_issue_events(events_res)
# }